refactor(cli): Phase A args split — common.rs + contract test
- Create args/ directory with mod.rs + common.rs
- Extract OutputFormat, ValidateOutputFormat, JudgeArgs to common.rs
- Add Cli::command().debug_assert() contract test
- Preserve crate::cli::args::* re-exports (no import churn)

See docs/contributing/SPLIT-PLAN-args-and-client.md
